refactor: rename backend files from task_revive to task_restore
- Rename task_revive.py to task_restore.py in endpoints and services
- Update imports in api.py to use task_restore module
