+++ backend/app/api/api.py
@@ -40,7 +40,7 @@ from app.api.endpoints.adapter import (
     subscriptions,
     task_knowledge_bases,
     task_members,
-    task_revive,
+    task_restore,
     tasks,
     teams,
 )
@@ -92,7 +92,7 @@ api_router.include_router(
     subscription_market.router, prefix="/market", tags=["subscription-market"]
 )
 api_router.include_router(tasks.router, prefix="/tasks", tags=["tasks"])
-api_router.include_router(task_revive.router, prefix="/tasks", tags=["tasks"])
+api_router.include_router(task_restore.router, prefix="/tasks", tags=["tasks"])
 api_router.include_router(subtasks.router, prefix="/subtasks", tags=["subtasks"])
 api_router.include_router(task_members.router, prefix="/tasks", tags=["task-members"])
 api_router.include_router(
